prac-02 solution

+++ DebuggingExercises/FixDebugTwo4.java
@@ -8,7 +8,7 @@ public class DebugTwo4
       final double tax = 0.06;
       costString = JOptionPane.showInputDialog(null,"Enter price of item you are buying", "Purchases",
               JOptionPane.INFORMATION_MESSAGE);
-      cost = Double.parseInt(costString);
+      cost = Double.parseDouble(costString);
       JOptionPane.showMessageDialog(null,"With " + (tax * 100) +
               "% tax,  purchase  is $" + (cost + cost * tax));
    }
